+++ BraintreeCore/src/main/java/com/braintreepayments/api/core/AnalyticsEvent.kt
@@ -26,5 +26,6 @@ internal data class AnalyticsEvent(
     val didEnablePayPalAppSwitch: Boolean? = null,
     val didPayPalServerAttemptAppSwitch: Boolean? = null,
     val didSdkAttemptAppSwitch: Boolean? = null,
-    val fundingSource: String? = null
+    val fundingSource: String? = null,
+    val uiType: String? = null
 )