+++ CHANGELOG.md
@@ -2,10 +2,19 @@
 
 ## unreleased
 
+* BraintreeCore
+    * Update Browser Switch version to 3.5.1
+* PayPal
+    * Add `ActivityResultRegistry` constructor to PayPalLauncher to support AuthTab integration from JetpackCompose
 * ShopperInsights
     * Add `PAYPAL_PAY_LATER` to `ButtonType` enum
 * GooglePay
     * Add optional `DisplayItems` in GooglePay request
+* UIComponents
+    * Add a compose `PayPalButton` class to generate a PayPal-branded button for launching PayPal flow
+    * Add a compose `VenmoButton` class to generate a Venmo-branded button for launching Venmo flow
+* Venmo
+    * Add `ActivityResultRegistry` constructor to VenmoLauncher to support AuthTab integration from JetpackCompose
 
 ## 5.24.0 (2026-02-25)